style: change update to event name
